refactor: extract helpers and validators, makes code readable (Jira: JSC-19)
